Fixed Table Formatting/Single-Cell Table Creation
check_cell.py was given a main() function so that tables can be created
by just running it (like they should be).
check_cell.py was renamed to single_cell_table.
single_cell_table.py
"Task" is now moved to the end of the print_table, so that the size of
the task name doesn't interfere with formatting.
Removed documentation of the cell warpiture, as it has been fixed.